added endpoint for current config, made some changes to directories and packages
