+++ bublight-backend/src/main/java/com/github/pyro2266/bublight/controllers/LedController.java
@@ -1,6 +1,6 @@
 package com.github.pyro2266.bublight.controllers;
 
-import com.github.pyro2266.bublight.modes.ColorModesProcessor;
+import com.github.pyro2266.bublight.service.colormodes.ColorModesService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,31 +18,31 @@ public class LedController {
 
     private static final Logger LOG = LoggerFactory.getLogger(LedController.class);
 
-    private final ColorModesProcessor colorModesProcessor;
+    private final ColorModesService colorModesService;
 
     @Autowired
-    public LedController(ColorModesProcessor colorModesProcessor) {
-        this.colorModesProcessor = colorModesProcessor;
+    public LedController(ColorModesService colorModesService) {
+        this.colorModesService = colorModesService;
     }
 
     @PostMapping(path = "/start")
     public ResponseEntity start() {
         LOG.info("Request to start color loop received");
-        colorModesProcessor.startColorLoop();
+        colorModesService.startColorLoop();
         return ResponseEntity.ok().build();
     }
 
     @PostMapping(path = "/stop")
     public ResponseEntity stop() {
         LOG.info("Request to stop color loop received");
-        colorModesProcessor.stopColorLoop();
+        colorModesService.stopColorLoop();
         return ResponseEntity.ok().build();
     }
 
     @PostMapping(path = "/setRefreshRate/{rate}")
     public ResponseEntity setRefreshRate(@PathVariable int rate) {
         LOG.info("Request to change refresh rate to {} received", rate);
-        colorModesProcessor.setRefreshRate(rate);
+        colorModesService.setRefreshRate(rate);
         return ResponseEntity.ok().build();
     }
 }
