+++ bublight-backend/src/main/java/com/github/pyro2266/bublight/controllers/LedModeController.java
@@ -1,14 +1,14 @@
 package com.github.pyro2266.bublight.controllers;
 
-import com.github.pyro2266.bublight.modes.ColorModesProcessor;
-import com.github.pyro2266.bublight.modes.base.simplecolor.SimpleColorMode;
-import com.github.pyro2266.bublight.modes.base.simplecolor.SimpleColorModeConfig;
-import com.github.pyro2266.bublight.modes.base.simplerainbow.SimpleRainbowMode;
-import com.github.pyro2266.bublight.modes.base.simplerainbow.SimpleRainbowModeConfig;
-import com.github.pyro2266.bublight.modes.overlay.brightnessbypressure.BrightnessByPressureMode;
-import com.github.pyro2266.bublight.modes.overlay.brightnessbypressure.BrightnessByPressureModeConfig;
-import com.github.pyro2266.bublight.modes.overlay.runningdot.RunningDotMode;
-import com.github.pyro2266.bublight.modes.overlay.runningdot.RunningDotModeConfig;
+import com.github.pyro2266.bublight.service.colormodes.ColorModesService;
+import com.github.pyro2266.bublight.service.colormodes.data.base.simplecolor.SimpleColorMode;
+import com.github.pyro2266.bublight.service.colormodes.data.base.simplecolor.SimpleColorModeConfig;
+import com.github.pyro2266.bublight.service.colormodes.data.base.simplerainbow.SimpleRainbowMode;
+import com.github.pyro2266.bublight.service.colormodes.data.base.simplerainbow.SimpleRainbowModeConfig;
+import com.github.pyro2266.bublight.service.colormodes.data.overlay.brightnessbypressure.BrightnessByPressureMode;
+import com.github.pyro2266.bublight.service.colormodes.data.overlay.brightnessbypressure.BrightnessByPressureModeConfig;
+import com.github.pyro2266.bublight.service.colormodes.data.overlay.runningdot.RunningDotMode;
+import com.github.pyro2266.bublight.service.colormodes.data.overlay.runningdot.RunningDotModeConfig;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,17 +27,17 @@ public class LedModeController {
 
     private static final Logger LOG = LoggerFactory.getLogger(LedModeController.class);
 
-    private final ColorModesProcessor colorModesProcessor;
+    private final ColorModesService colorModesService;
     private final SimpleColorMode simpleColorMode;
     private final SimpleRainbowMode simpleRainbowMode;
     private final BrightnessByPressureMode brightnessByPressureMode;
     private final RunningDotMode runningDotMode;
 
     @Autowired
-    public LedModeController(ColorModesProcessor colorModesProcessor, SimpleColorMode simpleColorMode,
-            SimpleRainbowMode simpleRainbowMode, BrightnessByPressureMode brightnessByPressureMode,
-            RunningDotMode runningDotMode) {
-        this.colorModesProcessor = colorModesProcessor;
+    public LedModeController(ColorModesService colorModesService, SimpleColorMode simpleColorMode,
+                             SimpleRainbowMode simpleRainbowMode, BrightnessByPressureMode brightnessByPressureMode,
+                             RunningDotMode runningDotMode) {
+        this.colorModesService = colorModesService;
         this.simpleColorMode = simpleColorMode;
         this.simpleRainbowMode = simpleRainbowMode;
         this.brightnessByPressureMode = brightnessByPressureMode;
@@ -48,7 +48,7 @@ public class LedModeController {
     public ResponseEntity activateSimpleColorMode(@RequestBody SimpleColorModeConfig config) {
         LOG.info("Activating {} with config {}", SimpleColorMode.class, config);
         simpleColorMode.setConfig(config);
-        colorModesProcessor.setBaseLedMode(simpleColorMode);
+        colorModesService.setBaseLedMode(simpleColorMode);
         return ResponseEntity.ok().build();
     }
 
@@ -63,7 +63,7 @@ public class LedModeController {
     public ResponseEntity activateSimpleRainbowMode(@RequestBody SimpleRainbowModeConfig config) {
         LOG.info("Activating {}", SimpleRainbowMode.class);
         simpleRainbowMode.setConfig(config);
-        colorModesProcessor.setBaseLedMode(simpleRainbowMode);
+        colorModesService.setBaseLedMode(simpleRainbowMode);
         return ResponseEntity.ok().build();
     }
 
@@ -78,7 +78,7 @@ public class LedModeController {
     public ResponseEntity activateBrightnessByPressureMode(@RequestBody BrightnessByPressureModeConfig config) {
         LOG.info("Activating {} with config {}", BrightnessByPressureMode.class, config);
         brightnessByPressureMode.setConfig(config);
-        colorModesProcessor.setOverlayLedMode(brightnessByPressureMode);
+        colorModesService.setOverlayLedMode(brightnessByPressureMode);
         return ResponseEntity.ok().build();
     }
 
@@ -93,7 +93,7 @@ public class LedModeController {
     public ResponseEntity activateRunningDotMode(@RequestBody RunningDotModeConfig config) {
         LOG.info("Activating {} with config {}", RunningDotMode.class, config);
         runningDotMode.setConfig(config);
-        colorModesProcessor.setOverlayLedMode(runningDotMode);
+        colorModesService.setOverlayLedMode(runningDotMode);
         return ResponseEntity.ok().build();
     }
 
@@ -107,7 +107,7 @@ public class LedModeController {
     @PostMapping(path = "/overlay/disable")
     public ResponseEntity activateRunningDotMode() {
         LOG.info("Disabling overlay mode");
-        colorModesProcessor.disableOverlayMode();
+        colorModesService.disableOverlayMode();
         return ResponseEntity.ok().build();
     }
 }
