--- bublight-backend/src/main/java/com/github/pyro2266/bublight/led/core/LedRendererService.java
@@ -1,14 +1,0 @@
-package com.github.pyro2266.bublight.led.core;
-
-import com.github.pyro2266.bublight.modes.Color;
-import org.springframework.stereotype.Service;
-
-@Service
-public interface LedRendererService {
-
-    /**
-     * Render colors on LED strip.
-     */
-    void renderColors(Color[] colors);
-
-}
