--- bublight-backend/src/main/java/com/github/pyro2266/bublight/modes/LedMode.java
@@ -1,8 +1,0 @@
-package com.github.pyro2266.bublight.modes;
-
-
-public interface LedMode {
-
-    String getModeId();
-
-}
