--- bublight-backend/src/main/java/com/github/pyro2266/bublight/modes/LedModeException.java
@@ -1,12 +1,0 @@
-package com.github.pyro2266.bublight.modes;
-
-public class LedModeException extends Exception {
-
-    public LedModeException(String s) {
-        super(s);
-    }
-
-    public LedModeException(String s, Throwable throwable) {
-        super(s, throwable);
-    }
-}
