--- bublight-backend/src/main/java/com/github/pyro2266/bublight/modes/OverlayLedMode.java
@@ -1,10 +1,0 @@
-package com.github.pyro2266.bublight.modes;
-
-public interface OverlayLedMode extends LedMode {
-
-    /**
-     * Modify base color configuration.
-     */
-    Color[] getNextColors(Color[] baseColors) throws LedModeException;
-
-}
