--- bublight-backend/src/main/java/com/github/pyro2266/bublight/pressure/PressureException.java
@@ -1,8 +1,0 @@
-package com.github.pyro2266.bublight.pressure;
-
-public class PressureException extends Exception {
-
-    public PressureException(String s, Throwable throwable) {
-        super(s, throwable);
-    }
-}
