--- bublight-backend/src/main/java/com/github/pyro2266/bublight/pressure/driver/PressureSensor.java
@@ -1,14 +1,0 @@
-package com.github.pyro2266.bublight.pressure.driver;
-
-import java.io.IOException;
-import org.springframework.stereotype.Service;
-
-@Service
-public interface PressureSensor {
-
-    /**
-     * Get current pressure
-     */
-    float readPressure() throws IOException, InterruptedException;
-
-}
