+++ bublight-backend/src/main/java/com/github/pyro2266/bublight/service/colormodes/data/LedMode.java
@@ -1,0 +1,8 @@
+package com.github.pyro2266.bublight.service.colormodes.data;
+
+
+public interface LedMode {
+
+    String getModeId();
+
+}
