+++ bublight-backend/src/main/java/com/github/pyro2266/bublight/service/colormodes/data/LedModeException.java
@@ -1,0 +1,12 @@
+package com.github.pyro2266.bublight.service.colormodes.data;
+
+public class LedModeException extends Exception {
+
+    public LedModeException(String s) {
+        super(s);
+    }
+
+    public LedModeException(String s, Throwable throwable) {
+        super(s, throwable);
+    }
+}
