+++ bublight-backend/src/main/java/com/github/pyro2266/bublight/service/config/ConfigService.java
@@ -1,0 +1,9 @@
+package com.github.pyro2266.bublight.service.config;
+
+import com.github.pyro2266.bublight.service.config.data.CurrentConfig;
+
+public interface ConfigService {
+
+    CurrentConfig getCurrentConfig();
+
+}
