+++ bublight-backend/src/main/java/com/github/pyro2266/bublight/service/pressure/data/PressureException.java
@@ -1,0 +1,8 @@
+package com.github.pyro2266.bublight.service.pressure.data;
+
+public class PressureException extends Exception {
+
+    public PressureException(String s, Throwable throwable) {
+        super(s, throwable);
+    }
+}
